+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 ## Changelog
 All notable changes to this project will be documented in this file.
 
+### 1.0.8
+#### Changeed
+  - Support debugger on weexSDK v0.18.+ (add sandbox feature).
+#### Fixed
+  - Fix native log can not be filter, [issues 355](https://github.com/weexteam/weex-toolkit/issues/355).
+
 ### 1.0.7
 #### Changeed
   - Formate time for logger.